Pass whole numbers to randrange in seed dispersal method 0
Dispersal method 0 (land anywhere in the world) called
random.randrange(-(theWorldSize/2), theWorldSize/2), which passes floats.
Python 3.12 and later refuse floats (TypeError), and 3.11 already refused
them when the world size is odd (ValueError, e.g. 10.5). This uses
int(theWorldSize/2), as placeSeed in vworldr.py already does. For worlds
of even size on Python 3.11 the numbers drawn are the same as before.

Tests: dispersal_methods_0_1_2 gains a stage with a world of size 21; its
first two stages are unchanged.

+++ Vida_Data/vplantr.py
@@ -297,8 +297,9 @@ class genericPlant(object):
         ###this dispersal method needs to be better
         if motherPlant.seedDispersalMethod[0]==0:
             ###This is just random anywhere in world###
-            newX =random.randrange(-(theGarden.theWorldSize/2),(theGarden.theWorldSize/2))+random.random()
-            newY =random.randrange(-(theGarden.theWorldSize/2),(theGarden.theWorldSize/2))+random.random()
+            #randrange needs whole numbers (the same as placeSeed in vworldr.py)
+            newX =random.randrange(-int(theGarden.theWorldSize/2),int(theGarden.theWorldSize/2))+random.random()
+            newY =random.randrange(-int(theGarden.theWorldSize/2),int(theGarden.theWorldSize/2))+random.random()
         elif motherPlant.seedDispersalMethod[0]==1:
             ###just drop the seed straight down###
             newX=theSeed.x
